tricky_SeqFn: ~(u64) on SHADOW_VISIBLE clear -> li -5; and (95.0->95.6%)
case 0x2b clears OBJ_MODEL_STATE_SHADOW_VISIBLE (0x4); the plain ~K emitted a
single-bit rlwinm where target materializes the inverse mask (li -5; and),
matching the sibling ~(u64)0x800 clears in this fn (#2/#74). Instr count T=C=292,
structural regions 3->2; residual = lfs scheduling reorder + r24/r25 coloring.

(cherry picked from commit 23662eb97c2ae63c61810107b4ae55ec6948c9f0)

+++ src/main/dll/dll_00C4_tricky.c
@@ -436,7 +436,7 @@ int tricky_SeqFn(int obj, int unused, ObjAnimUpdateState* animUpdate)
             **(u8**)&((TrickyState*)state)->progressPtr = ((TrickyState*)state)->unk82D;
             break;
         case 0x2b:
-            ((GameObject*)obj)->anim.modelState->flags &= ~OBJ_MODEL_STATE_SHADOW_VISIBLE;
+            ((GameObject*)obj)->anim.modelState->flags &= ~(u64)OBJ_MODEL_STATE_SHADOW_VISIBLE;
             break;
         case 0x2c:
             ((GameObject*)obj)->anim.modelState->flags |= OBJ_MODEL_STATE_SHADOW_VISIBLE;
